requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
idna==3.4
python-decouple==3.8
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0

# LangChain + Gemini
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
pydantic>=2.0.0
